+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/settings/UserSettingsViewModel.kt
@@ -14,7 +14,6 @@ interface UserSettingsViewModelFactory {
         onShowNotificationsSettings: () -> Unit,
         onShowPrivacySettings: () -> Unit,
         onShowAppearanceSettings: () -> Unit,
-        onShowAccountSetup: (userId: UserId) -> Unit,
     ): UserSettingsViewModel {
         return UserSettingsViewModelImpl(
             viewModelContext,
@@ -24,7 +23,6 @@ interface UserSettingsViewModelFactory {
             onShowNotificationsSettings,
             onShowPrivacySettings,
             onShowAppearanceSettings,
-            onShowAccountSetup,
         )
     }
 
@@ -38,7 +36,6 @@ interface UserSettingsViewModel {
     fun showNotificationsSettings()
     fun showPrivacySettings()
     fun showAppearanceSettings()
-    fun showAccountSetup(userId: UserId)
 }
 
 open class UserSettingsViewModelImpl(
@@ -49,7 +46,6 @@ open class UserSettingsViewModelImpl(
     private val onShowNotificationsSettings: () -> Unit,
     private val onShowPrivacySettings: () -> Unit,
     private val onShowAppearanceSettings: () -> Unit,
-    private val onShowAccountSetup: (userId : UserId) -> Unit,
 ) : ViewModelContext by viewModelContext, UserSettingsViewModel {
 
     private val backCallback = BackCallback {
@@ -82,9 +78,5 @@ open class UserSettingsViewModelImpl(
 
     override fun showAppearanceSettings() {
         onShowAppearanceSettings()
-    }
-
-    override fun showAccountSetup(userId: UserId) {
-        onShowAccountSetup(userId)
     }
 }